+++ examples/scenes/src/simple_text.rs
@@ -1,11 +1,11 @@
 // Copyright 2022 the Velato Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+use skrifa::raw::FontRef;
+use skrifa::MetadataProvider;
 use std::sync::Arc;
 use vello::kurbo::Affine;
 use vello::peniko::{Blob, Brush, BrushRef, Font, StyleRef};
-use vello::skrifa::raw::FontRef;
-use vello::skrifa::MetadataProvider;
 use vello::{Glyph, Scene};
 
 // This is very much a hack to get things working.
@@ -68,7 +68,7 @@ impl RobotoText {
         let brush = brush.into();
         let style = style.into();
         let axes = font_ref.axes();
-        let font_size = vello::skrifa::instance::Size::new(size);
+        let font_size = skrifa::instance::Size::new(size);
         let var_loc = axes.location(variations.iter().copied());
         let charmap = font_ref.charmap();
         let metrics = font_ref.metrics(font_size, &var_loc);
@@ -81,7 +81,7 @@ impl RobotoText {
             .font_size(size)
             .transform(transform)
             .glyph_transform(glyph_transform)
-            .normalized_coords(var_loc.coords())
+            .normalized_coords(bytemuck::cast_slice(var_loc.coords()))
             .brush(brush)
             .draw(
                 style,
@@ -129,7 +129,7 @@ impl RobotoText {
 }
 
 fn to_font_ref(font: &Font) -> Option<FontRef<'_>> {
-    use vello::skrifa::raw::FileRef;
+    use skrifa::raw::FileRef;
     let file_ref = FileRef::new(font.data.as_ref()).ok()?;
     match file_ref {
         FileRef::Font(font) => Some(font),