+++ examples/with_winit/src/stats.rs
@@ -43,7 +43,7 @@ impl Snapshot {
         scene.fill(
             Fill::NonZero,
             offset,
-            &Brush::Solid(Color::rgba8(0, 0, 0, 200)),
+            &Brush::Solid(Color::from_rgba8(0, 0, 0, 200)),
             None,
             &Rect::new(0., 0., width, height),
         );
@@ -130,9 +130,9 @@ impl Snapshot {
             let s = Affine::scale_non_uniform(1., -h);
             #[allow(clippy::match_overlapping_arm)]
             let color = match *sample {
-                ..=16_667 => Color::rgb8(100, 143, 255),
-                ..=33_334 => Color::rgb8(255, 176, 0),
-                _ => Color::rgb8(220, 38, 127),
+                ..=16_667 => Color::from_rgb8(100, 143, 255),
+                ..=33_334 => Color::from_rgb8(255, 176, 0),
+                _ => Color::from_rgb8(220, 38, 127),
             };
             scene.fill(
                 Fill::NonZero,