+++ src/runtime/model/mod.rs
@@ -15,33 +15,87 @@ pub use value::{Animated, Easing, EasingHandle, Time, Tween, Value, ValueRef};
 
 pub(crate) use spline::SplineToPath;
 
-macro_rules! simple_value {
-    ($name:ident) => {
-        #[allow(clippy::large_enum_variant)]
-        #[derive(Clone, Debug)]
-        pub enum $name {
-            Fixed(fixed::$name),
-            Animated(animated::$name),
+#[derive(Clone, Debug)]
+#[cfg_attr(
+    not(target_arch = "wasm32"),
+    expect(
+        clippy::large_enum_variant,
+        reason = "Deferred. Furthermore, for some reason, only on wasm32, this isn't triggering clippy."
+    )
+)]
+pub enum Transform {
+    Fixed(fixed::Transform),
+    Animated(animated::Transform),
+}
+impl Transform {
+    pub fn is_fixed(&self) -> bool {
+        matches!(self, Self::Fixed(_))
+    }
+    pub fn evaluate(&self, frame: f64) -> ValueRef<fixed::Transform> {
+        match self {
+            Self::Fixed(value) => ValueRef::Borrowed(value),
+            Self::Animated(value) => ValueRef::Owned(value.evaluate(frame)),
         }
-
-        impl $name {
-            pub fn is_fixed(&self) -> bool {
-                matches!(self, Self::Fixed(_))
-            }
-            pub fn evaluate(&self, frame: f64) -> ValueRef<fixed::$name> {
-                match self {
-                    Self::Fixed(value) => ValueRef::Borrowed(value),
-                    Self::Animated(value) => ValueRef::Owned(value.evaluate(frame)),
-                }
-            }
-        }
-    };
+    }
 }
 
-simple_value!(Transform);
-simple_value!(Stroke);
-simple_value!(Repeater);
-simple_value!(ColorStops);
+#[derive(Clone, Debug)]
+pub enum Stroke {
+    Fixed(fixed::Stroke),
+    Animated(animated::Stroke),
+}
+impl Stroke {
+    pub fn is_fixed(&self) -> bool {
+        matches!(self, Self::Fixed(_))
+    }
+    pub fn evaluate(&self, frame: f64) -> ValueRef<fixed::Stroke> {
+        match self {
+            Self::Fixed(value) => ValueRef::Borrowed(value),
+            Self::Animated(value) => ValueRef::Owned(value.evaluate(frame)),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+#[cfg_attr(
+    not(target_arch = "wasm32"),
+    expect(
+        clippy::large_enum_variant,
+        reason = "Deferred. Furthermore, for some reason, only on wasm32, this isn't triggering clippy."
+    )
+)]
+pub enum Repeater {
+    Fixed(fixed::Repeater),
+    Animated(animated::Repeater),
+}
+impl Repeater {
+    pub fn is_fixed(&self) -> bool {
+        matches!(self, Self::Fixed(_))
+    }
+    pub fn evaluate(&self, frame: f64) -> ValueRef<fixed::Repeater> {
+        match self {
+            Self::Fixed(value) => ValueRef::Borrowed(value),
+            Self::Animated(value) => ValueRef::Owned(value.evaluate(frame)),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum ColorStops {
+    Fixed(fixed::ColorStops),
+    Animated(animated::ColorStops),
+}
+impl ColorStops {
+    pub fn is_fixed(&self) -> bool {
+        matches!(self, Self::Fixed(_))
+    }
+    pub fn evaluate(&self, frame: f64) -> ValueRef<fixed::ColorStops> {
+        match self {
+            Self::Fixed(value) => ValueRef::Borrowed(value),
+            Self::Animated(value) => ValueRef::Owned(value.evaluate(frame)),
+        }
+    }
+}
 
 #[derive(Clone, Debug)]
 pub enum Brush {
@@ -60,7 +114,7 @@ impl Brush {
                 if alpha == 1.0 {
                     ValueRef::Borrowed(value)
                 } else {
-                    ValueRef::Owned(fixed::brush_with_alpha(value, alpha))
+                    ValueRef::Owned(value.to_owned().multiply_alpha(alpha as _))
                 }
             }
             Self::Animated(value) => ValueRef::Owned(value.evaluate(alpha, frame)),