+++ src/runtime/render.rs
@@ -66,7 +66,7 @@ impl Renderer {
         scene.pop_layer();
     }
 
-    #[allow(clippy::too_many_arguments)]
+    #[expect(clippy::too_many_arguments, reason = "Deferred")]
     fn render_layer(
         &mut self,
         animation: &Composition,
@@ -351,7 +351,7 @@ impl Batch {
         for draw in self.draws.iter().rev() {
             // Some nastiness to avoid cloning the brush if unnecessary
             let modified_brush = if draw.alpha != 1.0 {
-                Some(fixed::brush_with_alpha(&draw.brush, draw.alpha))
+                Some(draw.brush.clone().multiply_alpha(draw.alpha as _))
             } else {
                 None
             };