+++ src/schema/shapes/mod.rs
@@ -51,7 +51,13 @@ use serde::{Deserialize, Serialize};
 /// share the properties in `shapes::common::Properties`.
 #[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(tag = "ty")]
-#[allow(clippy::large_enum_variant)]
+#[cfg_attr(
+    not(target_arch = "wasm32"),
+    expect(
+        clippy::large_enum_variant,
+        reason = "Deferred. Furthermore, for some reason, only on wasm32, this isn't triggering clippy."
+    )
+)]
 pub enum AnyShape {
     /// A group is a shape that can contain other shapes (including other
     /// groups)